Report error when personals contain no ticket
When a class zephyr contains no ticket, it's probably a sign that somebody's
using -c adehnert for actual discussion or something. When a personal zephyr
contains no ticket, it's probably a sign that somebody forgotten to include the
tracker name or mispelled it or something. Rather than just ignore the zephyr,
we should let the user know so they don't just think Debothena is slow or
broken. Fixes Debothena-#12.