Remove superfluous is_personal definition
Back in 7019556, I added an is_personal function so that MIT class numbers in
personals would produce a response. By the time 726fef6a came around and I
wanted to disable the timeout for personals, I'd apparently forgotten, so I
added an identically-named equivalent function. This punts the newer one.